backend/requirements.txt
fastapi[standard]==0.128.4
uvicorn[standard]==0.19.0
websockets==10.3
python-deriv-api==0.1.6
pandas-ta
pandas
numpy
pydantic>=2.10
pydantic-settings
aiohttp
tweepy
sqlmodel
python-dotenv
